Sync from base fork

+++ src/net/sourceforge/plantuml/api/INumberAnalyzed.java
@@ -4,6 +4,11 @@
  *
  * Project Info:  http://plantuml.com
  * 
+ * If you like this project or if you find it useful, you can support us at:
+ * 
+ * http://plantuml.com/patreon (only 1$ per month!)
+ * http://plantuml.com/paypal
+ * 
  * This file is part of PlantUML.
  *
  * PlantUML is free software; you can redistribute it and/or modify it
@@ -25,14 +30,14 @@ package net.sourceforge.plantuml.api;
 
 public interface INumberAnalyzed {
 
-	public int getNb();
+	public long getNb();
 
-	public int getSum();
+	public long getSum();
 
-	public int getMin();
+	public long getMin();
 
-	public int getMax();
+	public long getMax();
 
-	public int getMean();
+	public long getMean();
 
 }
